Conflicts:
	archive/archive-core/src/main/java/dk/netarkivet/archive/bitarchive/BitarchiveAdmin.java
	archive/archive-core/src/main/resources/dk/netarkivet/archive/settings.xml
	common/common-core/pom.xml

+++ harvester/heritrix3/heritrix3-controller/src/main/java/dk/netarkivet/harvester/heritrix3/IngestableFiles.java
@@ -2,7 +2,7 @@
  * #%L
  * Netarchivesuite - harvester
  * %%
- * Copyright (C) 2005 - 2014 The Royal Danish Library, the Danish State and University Library,
+ * Copyright (C) 2005 - 2018 The Royal Danish Library, 
  *             the National Library of France and the Austrian National Library.
  * %%
  * This program is free software: you can redistribute it and/or modify
@@ -37,6 +37,8 @@ import dk.netarkivet.common.exceptions.IOFailure;
 import dk.netarkivet.common.exceptions.PermissionDenied;
 import dk.netarkivet.common.exceptions.IllegalState;
 import dk.netarkivet.common.utils.FileUtils;
+import dk.netarkivet.common.utils.Settings;
+import dk.netarkivet.harvester.HarvesterSettings;
 import dk.netarkivet.harvester.harvesting.metadata.MetadataFileWriter;
 
 /**
@@ -69,6 +71,8 @@ public class IngestableFiles {
 
     private String harvestnamePrefix;
 
+    public static final String METADATA_FILENAME_FORMAT = Settings.get(HarvesterSettings.METADATA_FILENAME_FORMAT);
+
     private Long harvestId;
 
     private File heritrixJobDir;	
@@ -95,7 +99,10 @@ public class IngestableFiles {
         // Create subdir 'metadata' if not already exists.
         FileUtils.createDir(getMetadataDir());
         // Create/scratch subdir 'tmp-meta'
-        FileUtils.removeRecursively(getTmpMetadataDir());
+        if (getTmpMetadataDir().isDirectory()) {
+        	FileUtils.removeRecursively(getTmpMetadataDir());
+        	log.warn("Removed directory {} with contents", getTmpMetadataDir());
+        }
         FileUtils.createDir(getTmpMetadataDir());
     }
 
@@ -121,30 +128,32 @@ public class IngestableFiles {
 
     /**
      * Marks generated metadata as final, closes the writer, and moves the temporary metadata file to its final
-     * position, if successful.
+     * position.
      *
-     * @param success True if metadata was successfully generated, false otherwise.
      * @throws PermissionDenied If the metadata has already been marked as ready, or if no metadata file exists upon
      * success.
      * @throws IOFailure if there is an error marking the metadata as ready.
      */
-    public void setMetadataGenerationSucceeded(boolean success) {
+    public void closeMetadataFile() {
         if (isMetadataReady()) {
             throw new PermissionDenied("Metadata file " + getMetadataFile().getAbsolutePath() + " already exists");
         }
-
-        if (success) {
-            writer.close(); // close writer down
-            if (!getTmpMetadataFile().exists()) {
-                String message = "No metadata was generated despite claims that metadata generation was successful.";
-                throw new PermissionDenied(message);
-            }
-            getTmpMetadataFile().renameTo(getMetadataFile());
-        } else {
-            error = true;
+        writer.close(); // close writer down
+        if (!getTmpMetadataFile().exists()) {
+            String message = "No metadata was generated despite claims that metadata generation was successful.";
+            throw new PermissionDenied(message);
         }
+        getTmpMetadataFile().renameTo(getMetadataFile());
     }
-
+    
+    /**
+     * Set error state. 
+     * @param isError True, if error, otherwise false;
+     */
+    public void setErrorState(boolean isError) {
+        error = isError;
+    }
+    
     /**
      * Get a MetaDatafileWriter for the temporary metadata file. Successive calls to this method on the same object will
      * return the same writer. Once the metadata have been finalized, calling this method will fail.
@@ -196,7 +205,7 @@ public class IngestableFiles {
      * @return metadata arc file as a File
      */
     protected File getMetadataFile() {
-        return new File(getMetadataDir(), MetadataFileWriter.getMetadataArchiveFileName(Long.toString(jobId)));
+        return new File(getMetadataDir(), MetadataFileWriter.getMetadataArchiveFileName(Long.toString(jobId), harvestId));
     }
 
     /**
@@ -214,7 +223,7 @@ public class IngestableFiles {
      * @return tmp-metadata arc file as a File
      */
     private File getTmpMetadataFile() {
-        return new File(getTmpMetadataDir(), MetadataFileWriter.getMetadataArchiveFileName(Long.toString(jobId)));
+        return new File(getTmpMetadataDir(), MetadataFileWriter.getMetadataArchiveFileName(Long.toString(jobId), harvestId));
     }
 
     /**
@@ -254,11 +263,7 @@ public class IngestableFiles {
     public File getReportsDir() {
         return new File(heritrixJobDir, "latest/reports");
     }
-    
-    
-    
-    
-    
+        
     /**
      * Get a list of all WARC files that should get ingested. Any open files should be closed with closeOpenFiles first.
      *
